refactor API models
